requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pyyaml==6.0
pillow==9.5.0
numpy==1.24.3
pillow
transformers
torch
torchvision
argparse
